Remove getters from CachingQRCodeGenerator only needed for tests
