+++ collect_app/src/main/java/org/odk/collect/android/preferences/qr/CachingQRCodeGenerator.java
@@ -2,21 +2,13 @@ package org.odk.collect.android.preferences.qr;
 
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
-import android.graphics.Color;
 
-import com.google.zxing.BarcodeFormat;
-import com.google.zxing.EncodeHintType;
 import com.google.zxing.WriterException;
-import com.google.zxing.common.BitMatrix;
-import com.google.zxing.qrcode.QRCodeWriter;
-import com.google.zxing.qrcode.decoder.ErrorCorrectionLevel;
 
-import org.odk.collect.android.R;
-import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.storage.StoragePathProvider;
 import org.odk.collect.android.storage.StorageSubdirectory;
-import org.odk.collect.android.utilities.CompressionUtils;
 import org.odk.collect.android.utilities.FileUtils;
+import org.odk.collect.android.utilities.QRCodeUtils;
 import org.odk.collect.android.utilities.SharedPreferencesUtils;
 import org.odk.collect.utilities.Consumer;
 
@@ -26,8 +18,6 @@ import java.security.MessageDigest;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
 
 import io.reactivex.Observable;
 import io.reactivex.android.schedulers.AndroidSchedulers;
@@ -37,8 +27,10 @@ import timber.log.Timber;
 import static org.odk.collect.android.preferences.AdminKeys.KEY_ADMIN_PW;
 import static org.odk.collect.android.preferences.GeneralKeys.KEY_PASSWORD;
 
-public class ObservableQRCodeGenerator implements QRCodeGenerator {
-    private static final int QR_CODE_SIDE_LENGTH = 400; // in pixels
+public class CachingQRCodeGenerator implements QRCodeGenerator {
+
+
+
     private static final String SETTINGS_MD5_FILE = ".collect-settings-hash";
 
     @Override
@@ -100,7 +92,7 @@ public class ObservableQRCodeGenerator implements QRCodeGenerator {
             // If the file is not found in the disk or md5Hash not matched
             if (bitmap == null) {
                 Timber.i("Generating QRCode...");
-                bitmap = generateQRBitMap(preferencesString, QR_CODE_SIDE_LENGTH);
+                bitmap = generateQRBitMap(preferencesString);
                 shouldWriteToDisk = true;
             }
 
@@ -123,38 +115,18 @@ public class ObservableQRCodeGenerator implements QRCodeGenerator {
         });
     }
 
-    @Override
-    public Bitmap generateQRBitMap(String data, int sideLength) throws IOException, WriterException {
+    private Bitmap generateQRBitMap(String data) throws IOException, WriterException {
         final long time = System.currentTimeMillis();
-        String compressedData = CompressionUtils.compress(data);
-
-        // Maximum capacity for QR Codes is 4,296 characters (Alphanumeric)
-        if (compressedData.length() > 4000) {
-            throw new IOException(Collect.getInstance().getString(R.string.encoding_max_limit));
-        }
-
-        Map<EncodeHintType, ErrorCorrectionLevel> hints = new HashMap<>();
-        hints.put(EncodeHintType.ERROR_CORRECTION, ErrorCorrectionLevel.L);
-        QRCodeWriter qrCodeWriter = new QRCodeWriter();
-        BitMatrix bitMatrix = qrCodeWriter.encode(compressedData, BarcodeFormat.QR_CODE, sideLength, sideLength, hints);
-
-        Bitmap bmp = Bitmap.createBitmap(sideLength, sideLength, Bitmap.Config.RGB_565);
-        for (int x = 0; x < sideLength; x++) {
-            for (int y = 0; y < sideLength; y++) {
-                bmp.setPixel(x, y, bitMatrix.get(x, y) ? Color.BLACK : Color.WHITE);
-            }
-        }
+        Bitmap bmp = QRCodeUtils.encode(data);
         Timber.i("QR Code generation took : %d ms", System.currentTimeMillis() - time);
         return bmp;
     }
 
-    @Override
-    public String getQRCodeFilepath() {
+    private String getQRCodeFilepath() {
         return new StoragePathProvider().getDirPath(StorageSubdirectory.SETTINGS) + File.separator + "collect-settings.png";
     }
 
-    @Override
-    public String getMd5CachePath() {
+    private String getMd5CachePath() {
         return new StoragePathProvider().getDirPath(StorageSubdirectory.SETTINGS) + File.separator + SETTINGS_MD5_FILE;
     }
 }
