+++ collect_app/src/main/java/org/odk/collect/android/preferences/qr/QRCodeGenerator.java
@@ -2,11 +2,8 @@ package org.odk.collect.android.preferences.qr;
 
 import android.graphics.Bitmap;
 
-import com.google.zxing.WriterException;
-
 import org.odk.collect.utilities.Consumer;
 
-import java.io.IOException;
 import java.util.Collection;
 
 import io.reactivex.Observable;
@@ -17,10 +14,4 @@ public interface QRCodeGenerator {
     void generateQRCode(Consumer<String> callback);
 
     Observable<Bitmap> generateQRCode(Collection<String> selectedPasswordKeys);
-
-    Bitmap generateQRBitMap(String data, int sideLength) throws IOException, WriterException;
-
-    String getQRCodeFilepath();
-
-    String getMd5CachePath();
 }
