+++ collect_app/src/main/java/org/odk/collect/android/utilities/QRCodeUtils.java
@@ -15,6 +15,7 @@
 package org.odk.collect.android.utilities;
 
 import android.graphics.Bitmap;
+import android.graphics.Color;
 
 import androidx.annotation.NonNull;
 
@@ -22,21 +23,33 @@ import com.google.zxing.BarcodeFormat;
 import com.google.zxing.BinaryBitmap;
 import com.google.zxing.ChecksumException;
 import com.google.zxing.DecodeHintType;
+import com.google.zxing.EncodeHintType;
 import com.google.zxing.FormatException;
 import com.google.zxing.LuminanceSource;
 import com.google.zxing.NotFoundException;
 import com.google.zxing.RGBLuminanceSource;
 import com.google.zxing.Reader;
 import com.google.zxing.Result;
+import com.google.zxing.WriterException;
+import com.google.zxing.common.BitMatrix;
 import com.google.zxing.common.HybridBinarizer;
 import com.google.zxing.multi.qrcode.QRCodeMultiReader;
+import com.google.zxing.qrcode.QRCodeWriter;
+import com.google.zxing.qrcode.decoder.ErrorCorrectionLevel;
+
+import org.odk.collect.android.R;
+import org.odk.collect.android.application.Collect;
 
 import java.io.IOException;
 import java.util.EnumMap;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.zip.DataFormatException;
 
 public class QRCodeUtils {
+
+    private static final int QR_CODE_SIDE_LENGTH = 400; // in pixels
+
     private QRCodeUtils() {
     }
 
@@ -49,6 +62,29 @@ public class QRCodeUtils {
         Reader reader = new QRCodeMultiReader();
         Result result = reader.decode(getBinaryBitmap(bitmap), tmpHintsMap);
         return CompressionUtils.decompress(result.getText());
+    }
+
+    public static Bitmap encode(String data) throws IOException, WriterException {
+        String compressedData = CompressionUtils.compress(data);
+
+        // Maximum capacity for QR Codes is 4,296 characters (Alphanumeric)
+        if (compressedData.length() > 4000) {
+            throw new IOException(Collect.getInstance().getString(R.string.encoding_max_limit));
+        }
+
+        Map<EncodeHintType, ErrorCorrectionLevel> hints = new HashMap<>();
+        hints.put(EncodeHintType.ERROR_CORRECTION, ErrorCorrectionLevel.L);
+        QRCodeWriter qrCodeWriter = new QRCodeWriter();
+        BitMatrix bitMatrix = qrCodeWriter.encode(compressedData, BarcodeFormat.QR_CODE, QR_CODE_SIDE_LENGTH, QR_CODE_SIDE_LENGTH, hints);
+
+        Bitmap bmp = Bitmap.createBitmap(QR_CODE_SIDE_LENGTH, QR_CODE_SIDE_LENGTH, Bitmap.Config.RGB_565);
+        for (int x = 0; x < QR_CODE_SIDE_LENGTH; x++) {
+            for (int y = 0; y < QR_CODE_SIDE_LENGTH; y++) {
+                bmp.setPixel(x, y, bitMatrix.get(x, y) ? Color.BLACK : Color.WHITE);
+            }
+        }
+
+        return bmp;
     }
 
     @NonNull
